refactor: Split population.py into population/ subpackage (base + archive)
