+++ src/saealib/population/__init__.py
@@ -1,0 +1,23 @@
+from saealib.population.base import (
+    Individual,
+    Population,
+    PopulationAttribute,
+    bind_property,
+    bind_property_array,
+)
+from saealib.population.archive import (
+    Archive,
+    ArchiveMixin,
+    ParetoArchive,
+    ParetoMixin,
+)
+
+__all__ = [
+    "Archive",
+    "ArchiveMixin",
+    "Individual",
+    "ParetoArchive",
+    "ParetoMixin",
+    "Population",
+    "PopulationAttribute",
+]
